Parent Class
This will introduce a parent class and make the previous classes a
child of it. This will avoid repetitive functions and make it easier to
maintain in the future. A directory containing shows and a directory
within a show, containing shots, are all directories and share common
operations like adding/removing content from them and retrieving basic
directory related information.

The parent class was originally titled CommonOperations but will be
renamed to DataRepository as the child classes are more specific data
data reposititories.