Type Hints, Rename and Bug Fix
Will add type hints throughout the script.

The parent class will be renamed to DataManager; this better reflects
the overall functionality of the base class as it does more than just hold
data.

Fix a bug where the showDatabase() and updateDatabase() method were not
working properly when called from an instance of the parent class.